reorganized a little
